feat: implement ProjectContainer and related components for project overview
- Added ProjectContainer component to manage project details and actions.
- Created ProjectHeader, ProjectStats, ProjectTabs, and ProjectContainerSkeleton for UI structure.
- Integrated InviteTeamModal and EditProjectModal for team management and project editing.
- Removed outdated BacklogView component to streamline project management interface.
- Updated EditTaskModal to handle task updates with new project context. #15
