+++ src/components/projects/Overview/ProjectContainerSkeleton.tsx
@@ -1,0 +1,71 @@
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle
+} from '@/components/ui/card'
+import { Skeleton } from '@/components/ui/skeleton'
+
+function ProjectContainerSkeleton() {
+  return (
+    <div className="h-full overflow-auto bg-gradient-to-br from-gray-50 to-blue-50/30">
+      <div className="p-8 max-w-7xl mx-auto space-y-6">
+        {/* Project Header Skeleton */}
+        <div className="flex items-center space-x-4">
+          <Skeleton className="h-12 w-12 rounded-full" />
+          <div className="flex-1">
+            <Skeleton className="h-6 w-40 mb-2" />
+            <Skeleton className="h-4 w-24" />
+          </div>
+        </div>
+
+        {/* Stats Cards Skeleton */}
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
+          {[1, 2, 3].map((i) => (
+            <Card className="border-0 shadow-lg" key={i}>
+              <CardHeader>
+                <Skeleton className="h-5 w-24 mb-2" />
+                <Skeleton className="h-4 w-16" />
+              </CardHeader>
+              <CardContent>
+                <Skeleton className="h-8 w-20" />
+              </CardContent>
+            </Card>
+          ))}
+        </div>
+
+        {/* Quick Actions Skeleton */}
+        <Card className="border-0 shadow-lg">
+          <CardHeader>
+            <Skeleton className="h-5 w-32 mb-2" />
+            <Skeleton className="h-4 w-40" />
+          </CardHeader>
+          <CardContent>
+            <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
+              {[1, 2, 3].map((i) => (
+                <Skeleton className="h-10 w-full rounded-md" key={i} />
+              ))}
+            </div>
+          </CardContent>
+        </Card>
+
+        {/* Tabs Skeleton */}
+        <div className="mt-6">
+          <div className="flex space-x-4 mb-4">
+            {[1, 2, 3].map((i) => (
+              <Skeleton className="h-8 w-24 rounded-md" key={i} />
+            ))}
+          </div>
+          <div className="space-y-4">
+            {[1, 2, 3].map((i) => (
+              <Skeleton className="h-32 w-full rounded-lg" key={i} />
+            ))}
+          </div>
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default ProjectContainerSkeleton
